shooteru: Change ownership of io_is_busy file to system.
Changing ownership to system allows CpuGovernorService
to modify io_is_busy based on votes from applications such
as the camcorder which need to turn of io_is_busy in
order to reduce the amount of time for which the CPU
runs at the maximum frequency. Without the ownership
change, CpuGovernorService would not be able to write
to the io_is_busy file.